requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anndata==0.11.1
array_api_compat==1.9.1
async-timeout==5.0.1
build==1.0.3
CacheControl==0.13.1
certifi==2022.12.7
charset-normalizer==3.3.2
chex==0.1.87
clarabel==0.9.0
cleo==2.1.0
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.3.0
crashtest==0.4.1
cryptography==44.0.0
cvxpy==1.6.0
cycler==0.12.1
decorator==4.4.2
distlib==0.3.8
docrep==0.3.2
dulwich==0.21.7
etils==1.11.0
fastjsonschema==2.19.1
filelock==3.16.1
flatbuffers==23.5.26
flax==0.10.2
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.2.0
get-annotations==0.1.2
graphviz==0.20.3
grpcio==1.68.1
h5py==3.12.1
humanfriendly==10.0
humanize==4.11.0
idna==3.4
igraph==0.11.8
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib_resources==6.4.5
installer==0.7.0
jaraco.classes==3.3.1
jax==0.4.37
jaxlib==0.4.36
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.0
kiwisolver==1.4.7
kneed==0.8.5
lazy_loader==0.3
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
ml_collections==1.0.0
ml_dtypes==0.5.0
mlflow>=2.10.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mudata==0.3.1
multidict==6.1.0
multipledispatch==1.0.0
natsort==8.4.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
numpyro==0.16.1
onnxruntime==1.17.0
opencv-python-headless==4.9.0.80
openTSNE==1.0.2
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.10.2
osqp==0.6.7.post3
pandas==2.2.3
patsy==1.0.1
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.24.1
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pooch==1.8.0
POT==0.9.5
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.1
py-pcha==0.1.3
PyMatting==1.1.12
pynndescent==0.5.13
pyparsing==3.1.4
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyro-api==0.1.2
pyro-ppl==1.9.1
pytorch-lightning==2.4.0
qdldl==0.1.7.post4
rapidfuzz==3.6.1
referencing==0.33.0
rembg==2.0.54
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.17.1
scanpy==1.10.4
scikit-image==0.22.0
scikit-learn==1.6.0
scikit-misc==0.5.1
scipy==1.14.1
scs==3.2.7
scvi-tools==1.2.2.post2
seaborn==0.13.2
SecretStorage==3.3.3
session_info==1.0.0
shellingham==1.5.4
simplejson==3.19.3
sparse==0.15.4
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorstore==0.1.69
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.1.30
tomli==2.0.1
tomlkit==0.12.3
toolz==1.0.0
tqdm==4.64.1
triton==2.2.0
trove-classifiers==2024.1.31
tzdata==2024.2
umap-learn==0.5.7
urllib3==1.26.13
virtualenv==20.25.0
webencodings==0.5.1
Werkzeug==3.1.3
xarray==2024.11.0
xlrd==2.0.1
yarl==1.18.3
zipp==3.21.0
faiss-cpu==1.10.0
hdbscan==0.8.40
louvain==0.8.2
